fix directory validation method and unit test
